+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- refactor(plugins): update plugin metadata to use author instead of company. (`eda9ddf`)
 - feat(plugins): enhance plugin signature verification and alert handling. (`38ffcd8`)
 
 ## 1.8.8 - 2026-06-25